requirements.txt
pyspark
apache-sedona
pytest
sqlparse
icecream
kafka-python